Add support for relative url/Jekyll baseurl
Reference to static files were absolute and did not support Jekyll's
`baseurl` parametre, which, if the website is deployed with a url that is
not root (e.g. `mirrors.tuna.tsinghua.edu.cn/mirrors`), a 404 will be
returned when fetching static files as they are not correctly
referenced, causing the website to display that javascript was not
supported by the browser.

To solve this issue, this Q&A is referenced:
https://talk.jekyllrb.com/t/relative-url-and-baseurl/2051,
by changing
`href="/static/114514.js"` to `"{{ site.baseurl }}/static/114514.js"`
we can eliminate the problem.

Nevertheless, it should be noted that there may possibly be other
static references which this single commit did not cover.